requirements.txt
google.generativeai
streamlit
python-dotenv
langchain
langchain_google_genai